big updateeeeeeeeeee &&& fix sml

+++ src/main/java/com/group19/OrderManagementSystem_backend/entity/Food.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import jakarta.persistence.Table;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -22,7 +23,7 @@ public class Food {
     private String foodName;
 
     @Column(name = "food_price")
-    private long foodPrice;
+    private BigDecimal foodPrice;
 
     @Column(name = "description")
     private String description;
